refacoring: create property converter interface and use converters for pdf and docx
