+++ Dexter.IndexStrategies/Property/FileToTextStrategy.cs
@@ -1,6 +1,7 @@
 ﻿using Dexter.Core.Interfaces;
 using Dexter.Core.Models.Config;
 using Dexter.Core.Models.IndexStrategy;
+using Dexter.IndexStrategies.Converters;
 using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Wordprocessing;
 using iTextSharp.text.pdf;
@@ -23,18 +24,18 @@ namespace Dexter.IndexStrategies.Property
         public void Execute(IndexFieldEvent e)
         {
             var filePath = HttpContext.Current.Server.MapPath(e.UmbracoProperty.Value.ToString());
-            var bytes = File.ReadAllBytes(filePath);
 
-            var text = "";
+            var text = string.Empty;
 
             switch(System.IO.Path.GetExtension(filePath))
             {
                 case ".pdf":
-                    text = ConvertPDFToText(bytes);
+                    text = new PdfToTextConverter().Convert(filePath);
                     break;
                 case ".doc":
+                    break;
                 case ".docx":
-                    text = ConvertWordToText(filePath);
+                    text = new DocxToTextConverter().Convert(filePath);
                     break;
             }
 
@@ -53,48 +54,6 @@ namespace Dexter.IndexStrategies.Property
                 for (var currentPageIndex = 1; currentPageIndex <= numberOfPages; currentPageIndex++)
                 {
                     sb.Append(PdfTextExtractor.GetTextFromPage(reader, currentPageIndex));
-                }
-            }
-            catch (Exception exception)
-            {
-                Console.WriteLine(exception.Message);
-            }
-
-            return sb.ToString();
-        }
-
-        private static string ConvertPDFToText(byte[] bytes)
-        {
-            var sb = new StringBuilder();
-
-            try
-            {
-                var reader = new PdfReader(bytes);
-                var numberOfPages = reader.NumberOfPages;
-
-                for (var currentPageIndex = 1; currentPageIndex <= numberOfPages; currentPageIndex++)
-                {
-                    sb.Append(PdfTextExtractor.GetTextFromPage(reader, currentPageIndex));
-                }
-            }
-            catch (Exception exception)
-            {
-                Console.WriteLine(exception.Message);
-            }
-
-            return sb.ToString();
-        }
-
-        private static string ConvertWordToText(string filePath)
-        {
-            var sb = new StringBuilder();
-
-            try
-            {
-                using (WordprocessingDocument wordDocument = WordprocessingDocument.Open(filePath, false))
-                {
-                    Body body = wordDocument.MainDocumentPart.Document.Body;
-                    sb.Append(body.InnerText);
                 }
             }
             catch (Exception exception)
